+++ src/lib/papa-parse.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import {Papa} from './papa';
+
+@NgModule({
+    imports: [],
+    providers: [
+        Papa
+    ]
+})
+export class PapaParseModule { }