+++ src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of papaparse
+ */
+
+export * from './lib/papa';
+export * from './lib/papa-parse.module';